refactor(cell_methods): split bandolier/vendor/interaction/world along feature seams
Pure code movement to bring four oversized cell-method files under the
700-line hard cap (CLAUDE.md "File organization"). Behavior-preserving;
no logic, signature, or wire-format changes. Re-export discipline keeps
every external `use` path identical — zero public-surface change.

inventory/bandolier.rs (893) -> bandolier/
  - active_slot.rs : flush_dirty_bandolier_ammo + handle_request_active_slot_change
                     (holster choreography, per-weapon ability grant)
  - ammo_change.rs : handle_request_ammo_change
  - mod.rs re-exports (pub flush, pub(crate) active-slot, pub(super) ammo-change)

player/vendor.rs (871) -> vendor/
  - wire.rs    : read_i32_array / read_trailing_template_id (+ tests)
  - session.rs : VendorSession / vendor_context / validate_template_id (+ tests)
  - train.rs   : handle_train_ability (+ tests)
  - mod.rs     : dispatch (train + purchase/sell/buyback/repair/recharge group)

player/interaction.rs (796) -> interaction/
  - interact.rs : the interact (right-click) handler — hostile reroute,
                  trainer open, tag/template/dialog fall-through
  - dialog.rs   : dialogButtonChoice (#479 open-dialog gate) + initialResponse
  - mod.rs      : dispatch + the interaction test suite (super::*)

player/world/mod.rs (794) -> new world/ submodules
  - auto_cycle.rs     : setAutoCycle toggle + persist_state_field_bits
  - system_options.rs : updateSystemOptions + parse_name_value_pairs
  - reload.rs         : handle_reload / maybe_trigger_reload_on_activate
                        + UNHOLSTER_DRAW_DURATION + ABILITY_RELOAD_WEAPON
  - item_sequence.rs  : fire_item_sequence
  - mod.rs slimmed to dispatch; existing tests.rs / system_options_tests.rs
    unchanged (reach moved items via mod.rs re-exports)
