--- crates/services/src/cell/cell_methods/player/vendor.rs
@@ -1,871 +1,0 @@
-use crate::cell::messages::CellToBaseMsg;
-use crate::cell::space_manager::SpaceManager;
-use tokio::sync::mpsc;
-
-use super::constants::*;
-
-fn read_i32_array(args: &[u8], offset: &mut usize) -> Option<Vec<(i32, i32)>> {
-    if args.len() < *offset + 4 {
-        return None;
-    }
-    let count = u32::from_le_bytes([
-        args[*offset],
-        args[*offset + 1],
-        args[*offset + 2],
-        args[*offset + 3],
-    ]) as usize;
-    *offset += 4;
-
-    // Bound count against remaining bytes BEFORE allocating: a malformed packet
-    // with count=u32::MAX would otherwise reserve ~32 GiB up-front and OOM.
-    let remaining = args.len().saturating_sub(*offset);
-    if count.saturating_mul(8) > remaining {
-        return None;
-    }
-
-    let mut items = Vec::with_capacity(count);
-    for _ in 0..count {
-        if args.len() < *offset + 8 {
-            return None;
-        }
-        let item_id = i32::from_le_bytes([
-            args[*offset],
-            args[*offset + 1],
-            args[*offset + 2],
-            args[*offset + 3],
-        ]);
-        let quantity = i32::from_le_bytes([
-            args[*offset + 4],
-            args[*offset + 5],
-            args[*offset + 6],
-            args[*offset + 7],
-        ]);
-        *offset += 8;
-        items.push((item_id, quantity));
-    }
-    Some(items)
-}
-
-fn read_trailing_template_id(args: &[u8], offset: usize) -> Option<i32> {
-    if args.len() >= offset + 4 {
-        Some(i32::from_le_bytes([
-            args[offset],
-            args[offset + 1],
-            args[offset + 2],
-            args[offset + 3],
-        ]))
-    } else {
-        None
-    }
-}
-
-/// Resolved vendor session for the player, looked up from server-side state.
-///
-/// `vendor_template_id` is read from the vendor entity itself rather than trusted
-/// from the client's request — this prevents a client from opening vendor A
-/// then submitting a purchase against vendor B's item lists.
-struct VendorSession {
-    player_id: i32,
-    vendor_entity_id: i32,
-    /// Server-side authoritative template id for the currently-open vendor.
-    /// May be `None` if the vendor entity has no template attached.
-    server_template_id: Option<i32>,
-}
-
-fn vendor_context(entity_id: u32, space_mgr: &SpaceManager) -> Option<VendorSession> {
-    let player = space_mgr.get_entity(entity_id)?;
-    let player_id = player.player_id?;
-    let vendor_entity_id_u32 = player.vendor_entity?;
-    let vendor_entity_id = vendor_entity_id_u32 as i32;
-    let server_template_id = space_mgr
-        .get_entity(vendor_entity_id_u32)
-        .and_then(|v| v.template_id);
-    Some(VendorSession {
-        player_id,
-        vendor_entity_id,
-        server_template_id,
-    })
-}
-
-/// Validate that the client-supplied template id matches the vendor that was
-/// opened. Returns the authoritative server-side id on success, or `None` after
-/// logging the mismatch.
-fn validate_template_id(
-    entity_id: u32,
-    op: &str,
-    session: &VendorSession,
-    client_template_id: i32,
-) -> Option<i32> {
-    match session.server_template_id {
-        Some(server_id) if server_id == client_template_id => Some(server_id),
-        Some(server_id) => {
-            tracing::warn!(
-                entity_id,
-                op,
-                server_template_id = server_id,
-                client_template_id,
-                vendor_entity_id = session.vendor_entity_id,
-                "vendor op rejected: client supplied template id does not match opened vendor"
-            );
-            None
-        }
-        None => {
-            tracing::warn!(
-                entity_id,
-                op,
-                client_template_id,
-                vendor_entity_id = session.vendor_entity_id,
-                "vendor op rejected: opened vendor has no template id (server cannot validate)"
-            );
-            None
-        }
-    }
-}
-
-pub async fn dispatch(
-    entity_id: u32,
-    method_index: u16,
-    args: &[u8],
-    tx: &mpsc::Sender<CellToBaseMsg>,
-    space_mgr: &mut SpaceManager,
-) -> bool {
-    match method_index {
-        TRAIN_ABILITY => {
-            if args.len() >= 4 {
-                let ability_id = i32::from_le_bytes([args[0], args[1], args[2], args[3]]);
-                handle_train_ability(entity_id, ability_id, tx, space_mgr).await;
-            } else {
-                tracing::warn!(
-                    entity_id,
-                    args_len = args.len(),
-                    "trainAbility: truncated args (need 4 bytes ability_id)"
-                );
-            }
-            true
-        }
-
-        PURCHASE_ITEMS | SELL_ITEMS | BUYBACK_ITEMS | REPAIR_ITEMS | RECHARGE_ITEMS => {
-            let op_name = match method_index {
-                PURCHASE_ITEMS => "purchaseItems",
-                SELL_ITEMS => "sellItems",
-                BUYBACK_ITEMS => "buybackItems",
-                REPAIR_ITEMS => "repairItems",
-                RECHARGE_ITEMS => "rechargeItems",
-                _ => unreachable!(),
-            };
-
-            let session = match vendor_context(entity_id, space_mgr) {
-                Some(s) => s,
-                None => {
-                    tracing::warn!(
-                        entity_id,
-                        op = op_name,
-                        "vendor op: no active vendor context (player_id or vendor_entity unset)"
-                    );
-                    return true;
-                }
-            };
-
-            let mut offset = 0;
-            let items = match read_i32_array(args, &mut offset) {
-                Some(items) => items,
-                None => {
-                    tracing::warn!(
-                        entity_id,
-                        op = op_name,
-                        args_len = args.len(),
-                        "vendor op: malformed item array in args"
-                    );
-                    return true;
-                }
-            };
-
-            let trailing_template_id = read_trailing_template_id(args, offset);
-
-            // For paid Repair/Recharge, `trailing_template_id` is optional (None
-            // signals "free repair"). For Purchase/Sell/Buyback, it's required;
-            // and in all cases where the client supplied one, it must match the
-            // vendor that was actually opened so a client can't spoof it.
-            let validated_template_id = match trailing_template_id {
-                Some(client_id) => {
-                    match validate_template_id(entity_id, op_name, &session, client_id) {
-                        Some(server_id) => Some(server_id),
-                        None => return true,
-                    }
-                }
-                None => None,
-            };
-
-            let msg = match method_index {
-                PURCHASE_ITEMS => match validated_template_id {
-                    Some(vendor_template_id) => CellToBaseMsg::PurchaseVendorItems {
-                        entity_id,
-                        player_id: session.player_id,
-                        vendor_entity_id: session.vendor_entity_id,
-                        vendor_template_id,
-                        items,
-                    },
-                    None => {
-                        tracing::warn!(
-                            entity_id,
-                            op = op_name,
-                            "vendor op: missing vendor_template_id"
-                        );
-                        return true;
-                    }
-                },
-                SELL_ITEMS => match validated_template_id {
-                    Some(vendor_template_id) => CellToBaseMsg::SellVendorItems {
-                        entity_id,
-                        player_id: session.player_id,
-                        vendor_entity_id: session.vendor_entity_id,
-                        vendor_template_id,
-                        items,
-                    },
-                    None => {
-                        tracing::warn!(
-                            entity_id,
-                            op = op_name,
-                            "vendor op: missing vendor_template_id"
-                        );
-                        return true;
-                    }
-                },
-                BUYBACK_ITEMS => match validated_template_id {
-                    Some(vendor_template_id) => CellToBaseMsg::BuybackVendorItems {
-                        entity_id,
-                        player_id: session.player_id,
-                        vendor_entity_id: session.vendor_entity_id,
-                        vendor_template_id,
-                        items,
-                    },
-                    None => {
-                        tracing::warn!(
-                            entity_id,
-                            op = op_name,
-                            "vendor op: missing vendor_template_id"
-                        );
-                        return true;
-                    }
-                },
-                REPAIR_ITEMS => CellToBaseMsg::RepairInventoryItems {
-                    entity_id,
-                    player_id: session.player_id,
-                    item_ids: items.iter().map(|(id, _)| *id).collect(),
-                    vendor_template_id: validated_template_id,
-                },
-                RECHARGE_ITEMS => CellToBaseMsg::RechargeInventoryItems {
-                    entity_id,
-                    player_id: session.player_id,
-                    item_ids: items.iter().map(|(id, _)| *id).collect(),
-                    vendor_template_id: validated_template_id,
-                },
-                _ => unreachable!(),
-            };
-
-            if let Err(e) = tx.send(msg).await {
-                tracing::warn!(
-                    entity_id,
-                    op = op_name,
-                    "vendor op: cell->base channel closed: {e}"
-                );
-            }
-            true
-        }
-
-        _ => false,
-    }
-}
-
-#[cfg(test)]
-mod read_i32_array_tests {
-    use super::read_i32_array;
-
-    #[test]
-    fn empty_array_returns_empty_vec() {
-        let buf = 0u32.to_le_bytes();
-        let mut off = 0;
-        assert_eq!(read_i32_array(&buf, &mut off), Some(vec![]));
-        assert_eq!(off, 4, "offset must advance past the count prefix");
-    }
-
-    #[test]
-    fn single_pair_parses_in_order() {
-        let mut buf = Vec::new();
-        buf.extend_from_slice(&1u32.to_le_bytes());
-        buf.extend_from_slice(&12345i32.to_le_bytes());
-        buf.extend_from_slice(&67i32.to_le_bytes());
-        let mut off = 0;
-        assert_eq!(read_i32_array(&buf, &mut off), Some(vec![(12345, 67)]));
-        assert_eq!(off, 12);
-    }
-
-    #[test]
-    fn multiple_pairs_preserve_input_order() {
-        let pairs: [(i32, i32); 3] = [(10, 1), (20, 2), (30, 3)];
-        let mut buf = Vec::new();
-        buf.extend_from_slice(&(pairs.len() as u32).to_le_bytes());
-        for (id, q) in &pairs {
-            buf.extend_from_slice(&id.to_le_bytes());
-            buf.extend_from_slice(&q.to_le_bytes());
-        }
-        let mut off = 0;
-        let parsed = read_i32_array(&buf, &mut off).expect("parse must succeed");
-        assert_eq!(parsed, pairs.to_vec());
-        assert_eq!(off, 4 + pairs.len() * 8);
-    }
-
-    #[test]
-    fn returns_none_when_too_short_for_count_prefix() {
-        let mut off = 0;
-        assert!(read_i32_array(&[0, 1, 2], &mut off).is_none());
-    }
-
-    #[test]
-    fn returns_none_when_payload_truncated_mid_pair() {
-        let mut buf = Vec::new();
-        buf.extend_from_slice(&2u32.to_le_bytes());
-        buf.extend_from_slice(&1i32.to_le_bytes());
-        buf.extend_from_slice(&1i32.to_le_bytes());
-        buf.extend_from_slice(&[0u8, 0u8]); // truncated 2nd pair
-        let mut off = 0;
-        assert!(read_i32_array(&buf, &mut off).is_none());
-    }
-
-    #[test]
-    fn rejects_oversized_count_without_allocating() {
-        // count = u32::MAX with only 4 bytes after the prefix would imply
-        // ~32 GiB of payload. The function must reject up-front so a
-        // malicious packet can't OOM the cell process via Vec::with_capacity.
-        let mut buf = Vec::new();
-        buf.extend_from_slice(&u32::MAX.to_le_bytes());
-        buf.extend_from_slice(&[0u8; 4]); // far less than u32::MAX * 8 bytes
-        let mut off = 0;
-        assert!(read_i32_array(&buf, &mut off).is_none());
-    }
-
-    #[test]
-    fn parses_at_nonzero_offset() {
-        // The function uses `*offset` for both bounds checks and indexing, so
-        // a non-zero starting offset must work the same as a fresh slice.
-        let mut buf = Vec::with_capacity(16);
-        buf.extend_from_slice(&[0xAA, 0xBB, 0xCC, 0xDD]); // junk prefix
-        buf.extend_from_slice(&1u32.to_le_bytes());
-        buf.extend_from_slice(&7i32.to_le_bytes());
-        buf.extend_from_slice(&8i32.to_le_bytes());
-        let mut off = 4;
-        assert_eq!(read_i32_array(&buf, &mut off), Some(vec![(7, 8)]));
-        assert_eq!(off, 16);
-    }
-}
-
-#[cfg(test)]
-mod vendor_context_tests {
-    use super::{vendor_context, VendorSession};
-    use crate::test_support::make_space_manager;
-
-    fn assert_session(
-        session: Option<VendorSession>,
-        expected_player_id: i32,
-        expected_vendor_entity_id: i32,
-        expected_template_id: Option<i32>,
-    ) {
-        let s = session.expect("vendor_context should return Some");
-        assert_eq!(s.player_id, expected_player_id);
-        assert_eq!(s.vendor_entity_id, expected_vendor_entity_id);
-        assert_eq!(s.server_template_id, expected_template_id);
-    }
-
-    #[test]
-    fn returns_none_when_entity_does_not_exist() {
-        let mgr = make_space_manager();
-        assert!(vendor_context(99999, &mgr).is_none());
-    }
-
-    #[test]
-    fn returns_none_when_player_id_missing() {
-        let mut mgr = make_space_manager();
-        mgr.create_entity(1, "Agnos", [0.0, 0.0, 0.0], [0.0; 3])
-            .unwrap();
-        // player.player_id deliberately left None.
-        let vendor = mgr.allocate_npc_id();
-        mgr.spawn_npc(vendor, "Agnos", [2.0, 0.0, 0.0], [0.0; 3])
-            .unwrap();
-        if let Some(p) = mgr.get_entity_mut(1) {
-            p.vendor_entity = Some(vendor);
-        }
-        assert!(vendor_context(1, &mgr).is_none());
-    }
-
-    #[test]
-    fn returns_none_when_vendor_entity_unset() {
-        let mut mgr = make_space_manager();
-        mgr.create_entity(1, "Agnos", [0.0, 0.0, 0.0], [0.0; 3])
-            .unwrap();
-        if let Some(p) = mgr.get_entity_mut(1) {
-            p.player_id = Some(4242);
-            // p.vendor_entity deliberately left None — no vendor opened.
-        }
-        assert!(vendor_context(1, &mgr).is_none());
-    }
-
-    #[test]
-    fn happy_path_returns_session_with_template_id() {
-        let mut mgr = make_space_manager();
-        mgr.create_entity(1, "Agnos", [0.0, 0.0, 0.0], [0.0; 3])
-            .unwrap();
-        let vendor = mgr.allocate_npc_id();
-        mgr.spawn_npc(vendor, "Agnos", [2.0, 0.0, 0.0], [0.0; 3])
-            .unwrap();
-        if let Some(v) = mgr.get_entity_mut(vendor) {
-            v.template_id = Some(9001);
-        }
-        if let Some(p) = mgr.get_entity_mut(1) {
-            p.player_id = Some(4242);
-            p.vendor_entity = Some(vendor);
-        }
-        assert_session(vendor_context(1, &mgr), 4242, vendor as i32, Some(9001));
-    }
-
-    #[test]
-    fn happy_path_returns_session_when_vendor_lacks_template_id() {
-        // The function reads template_id authoritatively from the vendor
-        // entity, falling back to None rather than fabricating a value. A
-        // missing template_id surfaces as `server_template_id: None` —
-        // validate_template_id at the caller treats that as a rejection
-        // (so a client can't open a templateless vendor and submit ops).
-        let mut mgr = make_space_manager();
-        mgr.create_entity(1, "Agnos", [0.0, 0.0, 0.0], [0.0; 3])
-            .unwrap();
-        let vendor = mgr.allocate_npc_id();
-        mgr.spawn_npc(vendor, "Agnos", [2.0, 0.0, 0.0], [0.0; 3])
-            .unwrap();
-        // Deliberately not setting template_id on the vendor.
-        if let Some(p) = mgr.get_entity_mut(1) {
-            p.player_id = Some(4242);
-            p.vendor_entity = Some(vendor);
-        }
-        assert_session(vendor_context(1, &mgr), 4242, vendor as i32, None);
-    }
-
-    #[test]
-    fn returns_session_with_no_template_when_vendor_entity_id_is_stale() {
-        // The player has a stale vendor_entity pointing at an id that doesn't
-        // exist (vendor despawned or never spawned). vendor_context still
-        // returns Some with server_template_id: None — the caller's
-        // validate_template_id arm logs and rejects the op.
-        let mut mgr = make_space_manager();
-        mgr.create_entity(1, "Agnos", [0.0, 0.0, 0.0], [0.0; 3])
-            .unwrap();
-        if let Some(p) = mgr.get_entity_mut(1) {
-            p.player_id = Some(4242);
-            p.vendor_entity = Some(123456); // no entity at this id
-        }
-        assert_session(vendor_context(1, &mgr), 4242, 123456, None);
-    }
-}
-
-/// Train an ability — full validation + base-side persistence + debit.
-///
-/// Cell-side validation:
-/// 1. Ability id exists in `space_mgr.ability_defs` (rejects typos)
-/// 2. Player has a `player_id` (no orphan-entity grants)
-/// 3. Player not already known the ability (no-op duplicate)
-/// 4. Ability is in player's `archetype_ability_tree` (no cross-class
-///    training)
-/// 5. Player level meets the tree entry's `level` requirement
-/// 6. Every `prerequisite_abilities` entry is in `entity.abilities`
-///
-/// On all checks passing, sends `CellToBaseMsg::TrainAbility` to the
-/// base. The base does the `training_points` debit + DB UPDATE +
-/// responds with `BaseToCellMsg::AbilityGranted`, which the cell's
-/// dispatcher handles (in `service/base_messages/mod.rs`) by adding
-/// the ability to `entity.abilities` and sending
-/// `onKnownAbilitiesUpdate` for the hotbar refresh.
-///
-/// For archetypes without a tree (Scientist/Asgard/Goa'uld/etc. — see
-/// Phase 7 content gap), step 4 fails immediately. This is the
-/// correct behavior: training is gated on content existing, and the
-/// player can still get starter abilities via `char_creation_abilities`
-/// (Phase 2).
-async fn handle_train_ability(
-    entity_id: u32,
-    ability_id: i32,
-    tx: &mpsc::Sender<CellToBaseMsg>,
-    space_mgr: &mut SpaceManager,
-) {
-    // Step 1: ability exists
-    if !space_mgr.ability_defs.contains_key(&ability_id) {
-        tracing::warn!(
-            entity_id,
-            ability_id,
-            "trainAbility: ability_id not found in ability_defs — rejecting"
-        );
-        return;
-    }
-
-    // Snapshot the player's state for validation; no mutation here.
-    let (player_id, archetype_id, player_level, already_known) = {
-        let entity = match space_mgr.get_entity(entity_id) {
-            Some(e) => e,
-            None => return,
-        };
-        (
-            entity.player_id,
-            entity.archetype_id,
-            entity.level as i32,
-            entity.abilities.has_ability(ability_id),
-        )
-    };
-
-    // Step 2: player must have a player_id (rules out orphan entities,
-    // GM grants on NPCs, etc.)
-    let player_id = match player_id {
-        Some(pid) => pid,
-        None => {
-            tracing::warn!(
-                entity_id,
-                ability_id,
-                "trainAbility: entity has no player_id — rejecting"
-            );
-            return;
-        }
-    };
-
-    // Step 3: already-known is a silent no-op (replayed packet, UI
-    // double-click)
-    if already_known {
-        tracing::debug!(
-            entity_id,
-            player_id,
-            ability_id,
-            "trainAbility: ability already known — no-op"
-        );
-        return;
-    }
-
-    // Step 4: ability must be in the player's archetype training tree
-    let archetype_id = match archetype_id {
-        Some(a) => a,
-        None => {
-            tracing::warn!(
-                entity_id,
-                player_id,
-                ability_id,
-                "trainAbility: entity has no archetype_id — rejecting"
-            );
-            return;
-        }
-    };
-    let tree_entry = space_mgr
-        .archetype_ability_trees
-        .get(&archetype_id)
-        .and_then(|tree| tree.iter().find(|e| e.ability_id == ability_id));
-    let tree_entry = match tree_entry {
-        Some(e) => e,
-        None => {
-            tracing::info!(
-                target: "abilities",
-                event = "train_rejected",
-                reason = "not_in_archetype_tree",
-                entity_id,
-                player_id,
-                archetype_id,
-                ability_id,
-                "trainAbility: ability not in player's archetype tree — rejecting \
-                 (likely an unsupported archetype until Phase 7 content lands)"
-            );
-            return;
-        }
-    };
-
-    // Step 5: level requirement
-    if player_level < tree_entry.level {
-        tracing::info!(
-            target: "abilities",
-            event = "train_rejected",
-            reason = "level_too_low",
-            entity_id,
-            player_id,
-            ability_id,
-            player_level,
-            required_level = tree_entry.level,
-            "trainAbility: player level below required — rejecting"
-        );
-        return;
-    }
-
-    // Step 6: prerequisites
-    let prereqs = tree_entry.prerequisite_abilities.clone();
-    let missing_prereq = {
-        let entity = match space_mgr.get_entity(entity_id) {
-            Some(e) => e,
-            None => return,
-        };
-        prereqs
-            .iter()
-            .find(|&&pid| !entity.abilities.has_ability(pid))
-            .copied()
-    };
-    if let Some(missing) = missing_prereq {
-        tracing::info!(
-            target: "abilities",
-            event = "train_rejected",
-            reason = "missing_prerequisite",
-            entity_id,
-            player_id,
-            ability_id,
-            missing_prereq = missing,
-            "trainAbility: prerequisite ability not known — rejecting"
-        );
-        return;
-    }
-
-    // All validation passed. Hand off to base for the training_points
-    // debit + DB persist. The cell-side state change happens when base
-    // responds with AbilityGranted (see service/base_messages/mod.rs).
-    tracing::info!(
-        target: "abilities",
-        event = "train_requested",
-        entity_id,
-        player_id,
-        ability_id,
-        archetype_id,
-        "trainAbility: validation passed, requesting base persist + debit"
-    );
-    if let Err(e) = tx
-        .send(CellToBaseMsg::TrainAbility {
-            entity_id,
-            player_id,
-            ability_id,
-        })
-        .await
-    {
-        // Cell→base channel closed mid-train. The player saw their
-        // request validated client-side (training UI accepted the click)
-        // but the persistence step never ran. Log at error so operators
-        // can correlate against the "Train clicked but TP didn't drop"
-        // player report and investigate the channel-shutdown root cause.
-        tracing::error!(
-            target: "abilities",
-            event = "train_ability_send_failed",
-            entity_id,
-            player_id,
-            ability_id,
-            error = %e,
-            "TrainAbility cell→base send failed — training point not debited"
-        );
-    }
-}
-
-#[cfg(test)]
-mod handle_train_ability_tests {
-    //! Validation guards on `handle_train_ability` (cell side). The
-    //! base-side persist + debit is exercised via the live-DB path; these
-    //! unit tests pin the rejection cases so a refactor can't silently
-    //! drop a guard.
-    use super::*;
-    use crate::cell::messages::CellToBaseMsg;
-    use crate::cell::space_manager::SpaceManager;
-    use crate::cell::spawner::ArchetypeAbilityTreeEntry;
-    use cimmeria_entity::abilities::AbilityDef;
-    use tokio::sync::mpsc;
-
-    const TEST_ABILITY: i32 = 12345;
-
-    fn make_mgr() -> SpaceManager {
-        let mut mgr = SpaceManager::new(1);
-        let xml = r#"<?xml version="1.0"?><Spaces><Space WorldName="W" Instanced="false" MinX="-50" MaxX="50" MinY="-50" MaxY="50" /></Spaces>"#;
-        mgr.parse_spaces_xml(xml).unwrap();
-        mgr.create_startup_spaces(
-            r#"<?xml version="1.0"?><Spaces><Space WorldName="W" /></Spaces>"#,
-        )
-        .unwrap();
-        mgr.create_entity(1, "W", [0.0; 3], [0.0; 3]).unwrap();
-        mgr
-    }
-
-    fn seed_ability(mgr: &mut SpaceManager, ability_id: i32) {
-        mgr.ability_defs.insert(
-            ability_id,
-            AbilityDef {
-                ability_id,
-                name: "TestAbility".to_string(),
-                cooldown: 0.0,
-                warmup: 0.0,
-                flags: 0,
-                is_ranged: false,
-                min_range: 0,
-                max_range: 0,
-                target_type_id: 0,
-                effect_ids: vec![],
-                moniker_ids: vec![],
-                required_ammo: 0,
-                event_set_id: None,
-                velocity: 0.0,
-            },
-        );
-    }
-
-    fn seed_tree(mgr: &mut SpaceManager, archetype_id: i32, entry: ArchetypeAbilityTreeEntry) {
-        mgr.archetype_ability_trees
-            .entry(archetype_id)
-            .or_default()
-            .push(entry);
-    }
-
-    fn drain_train_msgs(rx: &mut mpsc::Receiver<CellToBaseMsg>) -> Vec<i32> {
-        let mut out = Vec::new();
-        while let Ok(msg) = rx.try_recv() {
-            if let CellToBaseMsg::TrainAbility { ability_id, .. } = msg {
-                out.push(ability_id);
-            }
-        }
-        out
-    }
-
-    #[tokio::test]
-    async fn rejects_unknown_ability_id() {
-        // No seeded ability_def → step 1 fails.
-        let mut mgr = make_mgr();
-        let (tx, mut rx) = mpsc::channel(8);
-        handle_train_ability(1, 99999, &tx, &mut mgr).await;
-        assert!(
-            drain_train_msgs(&mut rx).is_empty(),
-            "must not send TrainAbility"
-        );
-    }
-
-    #[tokio::test]
-    async fn rejects_entity_without_player_id() {
-        let mut mgr = make_mgr();
-        seed_ability(&mut mgr, TEST_ABILITY);
-        // Entity 1 has no player_id by default.
-        let (tx, mut rx) = mpsc::channel(8);
-        handle_train_ability(1, TEST_ABILITY, &tx, &mut mgr).await;
-        assert!(
-            drain_train_msgs(&mut rx).is_empty(),
-            "no player_id = no train"
-        );
-    }
-
-    #[tokio::test]
-    async fn already_known_ability_is_silent_noop() {
-        let mut mgr = make_mgr();
-        seed_ability(&mut mgr, TEST_ABILITY);
-        if let Some(e) = mgr.get_entity_mut(1) {
-            e.player_id = Some(100);
-            e.archetype_id = Some(1); // Soldier
-            e.level = 1;
-            e.abilities.add_ability(TEST_ABILITY); // already known
-        }
-        let (tx, mut rx) = mpsc::channel(8);
-        handle_train_ability(1, TEST_ABILITY, &tx, &mut mgr).await;
-        assert!(
-            drain_train_msgs(&mut rx).is_empty(),
-            "duplicate train is no-op"
-        );
-    }
-
-    #[tokio::test]
-    async fn rejects_ability_not_in_archetype_tree() {
-        // Player is a Soldier; ability is in a tree but only for Commando.
-        let mut mgr = make_mgr();
-        seed_ability(&mut mgr, TEST_ABILITY);
-        seed_tree(
-            &mut mgr,
-            2, // Commando
-            ArchetypeAbilityTreeEntry {
-                ability_id: TEST_ABILITY,
-                tree_index: 0,
-                level: 1,
-                prerequisite_abilities: vec![],
-            },
-        );
-        if let Some(e) = mgr.get_entity_mut(1) {
-            e.player_id = Some(100);
-            e.archetype_id = Some(1); // Soldier, NOT Commando
-            e.level = 1;
-        }
-        let (tx, mut rx) = mpsc::channel(8);
-        handle_train_ability(1, TEST_ABILITY, &tx, &mut mgr).await;
-        assert!(
-            drain_train_msgs(&mut rx).is_empty(),
-            "cross-class training rejected"
-        );
-    }
-
-    #[tokio::test]
-    async fn rejects_below_required_level() {
-        let mut mgr = make_mgr();
-        seed_ability(&mut mgr, TEST_ABILITY);
-        seed_tree(
-            &mut mgr,
-            1,
-            ArchetypeAbilityTreeEntry {
-                ability_id: TEST_ABILITY,
-                tree_index: 0,
-                level: 10, // requires level 10
-                prerequisite_abilities: vec![],
-            },
-        );
-        if let Some(e) = mgr.get_entity_mut(1) {
-            e.player_id = Some(100);
-            e.archetype_id = Some(1);
-            e.level = 5; // below requirement
-        }
-        let (tx, mut rx) = mpsc::channel(8);
-        handle_train_ability(1, TEST_ABILITY, &tx, &mut mgr).await;
-        assert!(drain_train_msgs(&mut rx).is_empty(), "level-gated reject");
-    }
-
-    #[tokio::test]
-    async fn rejects_missing_prerequisite_ability() {
-        let mut mgr = make_mgr();
-        seed_ability(&mut mgr, TEST_ABILITY);
-        seed_tree(
-            &mut mgr,
-            1,
-            ArchetypeAbilityTreeEntry {
-                ability_id: TEST_ABILITY,
-                tree_index: 0,
-                level: 1,
-                prerequisite_abilities: vec![1000], // requires ability 1000
-            },
-        );
-        if let Some(e) = mgr.get_entity_mut(1) {
-            e.player_id = Some(100);
-            e.archetype_id = Some(1);
-            e.level = 1;
-            // Doesn't know prerequisite 1000.
-        }
-        let (tx, mut rx) = mpsc::channel(8);
-        handle_train_ability(1, TEST_ABILITY, &tx, &mut mgr).await;
-        assert!(drain_train_msgs(&mut rx).is_empty(), "prereq-gated reject");
-    }
-
-    #[tokio::test]
-    async fn passes_all_guards_sends_train_ability_to_base() {
-        let mut mgr = make_mgr();
-        seed_ability(&mut mgr, TEST_ABILITY);
-        seed_tree(
-            &mut mgr,
-            1,
-            ArchetypeAbilityTreeEntry {
-                ability_id: TEST_ABILITY,
-                tree_index: 0,
-                level: 1,
-                prerequisite_abilities: vec![],
-            },
-        );
-        if let Some(e) = mgr.get_entity_mut(1) {
-            e.player_id = Some(100);
-            e.archetype_id = Some(1);
-            e.level = 5;
-        }
-        let (tx, mut rx) = mpsc::channel(8);
-        handle_train_ability(1, TEST_ABILITY, &tx, &mut mgr).await;
-        let sent = drain_train_msgs(&mut rx);
-        assert_eq!(sent, vec![TEST_ABILITY], "valid train forwarded to base");
-    }
-}
